model update

+++ src/com/luxoft/bankapp/exceptions/OverDraftLimitExceededException.java
@@ -5,15 +5,14 @@ public class OverDraftLimitExceededException extends NotEnoughFundsException {
 	private static final long serialVersionUID = 1L;
 	private String account;
 
-	public OverDraftLimitExceededException(String account, float amount) {
+	public OverDraftLimitExceededException(String account, double amount) {
 		super(amount);
 		this.account = account;
 	}
 
 	@Override
 	public String getMessage() {
-		return "Overdraft Limit exceeded on " + this.account + " amount: "
-				+ amount;
+		return "Overdraft Limit exceeded on " + this.account + " amount: " + amount;
 
 	}
 }
